LangChain Server/requirements.txt
fastapi==0.103.2
openai==0.27.8
python-dotenv==1.0.0
langchain==0.0.308
pydantic==2.4.2
click==8.1.7
uvicorn==0.24.0
PyYAML
websockets